# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Ikaros Management API
  version: 1.0.0
  description: |
    Especificación OpenAPI de la API del backend Laravel para Ikaros Management 2.
    Esta documentación refleja los endpoints actualmente definidos en la ruta principal de la API.
servers:
  - url: http://localhost:8000/api
    description: Backend local de desarrollo
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /login:
    post:
      tags: [Auth]
      summary: Inicia sesión de usuario.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                company:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Login correcto
...
  /users:
    get:
      tags: [Users]
      summary: Lista usuarios.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuarios encontrados
    post:
      tags: [Users]
      summary: Crea un usuario.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Usuario creado
  /users/{id}:
    get:
      tags: [Users]
      summary: Obtiene un usuario por ID.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usuario
    put:
      tags: [Users]
      summary: Actualiza un usuario.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Usuario actualizado
    delete:
      tags: [Users]
      summary: Elimina un usuario.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usuario eliminado
...
  /guardaPermisos:
    post:
      tags: [Roles]
      summary: Guarda permisos asociados a un rol.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Permisos guardados
  /roles/{idRol}/permisos:
    get:
      tags: [Roles]
      summary: Obtiene permisos de un rol.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: idRol
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Permisos del rol
...
  /uploaderFile:
    post:
      tags: [Files]
      summary: Sube un archivo al sistema.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Archivo subido
...
  /downloadFile/{uuid}/{folder}/{file}:
    get:
      tags: [Files]
      summary: Descarga un archivo por UUID, carpeta y nombre.
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
        - in: path
          name: folder
          required: true
          schema:
            type: string
        - in: path
          name: file
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Archivo descargado
...
  /checkIMAP:
    post:
      tags: [Mail]
      summary: Verifica la configuración IMAP del tenant.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Resultado del chequeo IMAP
...